Add files via upload
This Jupyter Notebook is the complete Python script for this project. It begins with data preprocessing, and balancing the dataset to prevent bias. It then constructs a neural network model that integrates BERT for text encoding, utilizing TensorFlow and Keras. The training process is detailed, including the use of metrics such as accuracy, precision, and recall to evaluate the model's performance. The notebook also includes visualizations of the performance metrics, such as a confusion matrix and a classification report, demonstrating the model's ability to accurately distinguish between spam and legitimate emails. This code serves as a practical implementation guide for using BERT in spam email classification tasks.